Merge MIT to Staging

+++ brd-ios/scripts/cosmos_apple_silicon.sh
@@ -1,5 +1,12 @@
 #!/bin/zsh
 
+if [[ $(uname -m) != 'arm64' ]]; then
+    echo "Running on X86"
+    exit 0
+fi
+
+echo "Running on ARM"
+
 ARM="ios-arm64"
 COSMOS="Cosmos"
 
@@ -20,6 +27,14 @@ ARM_SIMULATOR_PATH="$FRAMEWORK_PATH/$ARM_SIMULATOR"
 X86_BINARY_PATH="$FRAMEWORK_PATH/$X86_SIMULATOR/$COSMOS_FRAMEWORK/$COSMOS"
 ARM_BINARY_PATH="$FRAMEWORK_PATH/$ARM_SIMULATOR/$COSMOS_FRAMEWORK/$COSMOS"
 FAT_BINARY_PATH="$FRAMEWORK_PATH/$FAT_SIMULATOR/"
+
+MODIFICATION_INDICATOR="modified_cosmos_dir_do_not_delete.txt"
+
+if [ -f $FRAMEWORK_PATH/$MODIFICATION_INDICATOR ]; then
+    echo "Framework has already been modified for ARM devices."
+    exit 0
+fi
+
 
 cp -r $FRAMEWORK_PATH/$ARM $ARM_SIMULATOR_PATH
 
@@ -43,3 +58,5 @@ lipo -create -output "$FAT_BINARY_PATH/$COSMOS_FRAMEWORK/$COSMOS" $ARM_BINARY_PA
 rm -r -f "$FRAMEWORK_PATH/$X86_SIMULATOR/"
 rm -r -f "$ARM_SIMULATOR_PATH"
 mv "$FAT_BINARY_PATH" "$FRAMEWORK_PATH/$X86_SIMULATOR"
+
+touch $FRAMEWORK_PATH/$MODIFICATION_INDICATOR
